requirements.txt
python-nmap==0.6.4
PyYAML==5.4.1
dnspython==2.1.0
